pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hello-world-api"
version = "1.0.0"
description = "A simple Hello World REST API built with FastAPI"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "ASP Code Agent", email = "agent@example.com"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Software Development :: Libraries :: Application Frameworks"
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "sqlalchemy==2.0.44",
    "alembic==1.17.1",
    "psycopg2-binary==2.9.9",
    "asyncpg==0.29.0",
    "passlib[bcrypt]==1.7.4",
    "httpx==0.28.1",
    "requests==2.32.5",
    "python-dotenv==1.2.1",
    "python-dateutil==2.9.0.post0",
    "structlog==23.2.0",
    "rich==13.9.4",
    "langfuse",
    "anthropic",
    "tenacity",
    "python-fasthtml"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.2",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "isort>=5.12.0",
    "mypy>=1.7.1",
    "ruff>=0.1.0",
    "pylint>=3.0.0",
    "pytest-mock==3.12.0",
    "pdb-attach",
    "py-spy==0.3.14",
    "sphinx==7.2.6",
    "sphinx-rtd-theme==1.3.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0"
]

[project.urls]
Homepage = "https://github.com/example/hello-world-api"
Repository = "https://github.com/example/hello-world-api.git"
Issues = "https://github.com/example/hello-world-api/issues"
Documentation = "https://github.com/example/hello-world-api#readme"

# Note: Demo FastAPI app moved to artifacts/demo_hello_world_api.py
# No project scripts defined for the ASP platform itself

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | artifacts
)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
extend_skip = ["artifacts"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.coverage.run]
source = ["main"]
omit = [
    "tests/*",
    "venv/*",
    "*/venv/*",
    "*/site-packages/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod"
]

[tool.ruff]
line-length = 88
target-version = "py312"
exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    "venv",
    "build",
    "dist",
    "__pycache__",
    "artifacts",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports
"tests/**/*.py" = ["ARG", "S101"]  # unused arguments and assert usage in tests
